backend/requirements.txt
hashlib
singlestoredb
python-dotenv
google-cloud-storage
flask-cors
flask
